Refactor calendar account to import package
